395-time-series-rdb-connection: removed some duplicate tests, moved some undeprecated constructor tests to TimeSeriesClientTest, set MockConnection object in the WithoutConn tests
